More uncomplets

+++ LW1/Models/Airport.cs
@@ -8,7 +8,7 @@
         public int Runways { get; set; }
         public int SoldTickets { get; set; }
         public double AveragePassengers { get; set; }
-        public int MonthlyIncome { get; set; }
+        public double MonthlyIncome { get; set; }
         public int IncidentsCount { get; set; }
 
         public Airport()
