+++ LW1/Models/Db/DbAirport.cs
@@ -8,6 +8,6 @@ public class DbAirport
     public int Runways { get; set; }
     public int SoldTickets { get; set; }
     public double AveragePassengers { get; set; }
-    public int MonthlyIncome { get; set; }
+    public double MonthlyIncome { get; set; }
     public int IncidentsCount { get; set; }
 }
